+++ test/test-applications/integrations/TestApplication.Plugins/Plugin.OpAmp.cs
@@ -3,7 +3,8 @@
 
 using OpenTelemetry.AutoInstrumentation.PluginApi;
 using OpenTelemetry.AutoInstrumentation.PluginApi.OpAmp;
-using OpenTelemetry.OpAmp.Client;
+using OpenTelemetry.OpAmp.Client.Listeners;
+using OpenTelemetry.OpAmp.Client.Messages;
 using OpenTelemetry.OpAmp.Client.Settings;
 
 namespace TestApplication.Plugins;
@@ -12,9 +13,11 @@ namespace TestApplication.Plugins;
 /// <summary>
 /// OpAMP extensions of the plugin.
 /// </summary>
-public partial class Plugin : IPlugin, IOpAmpPlugin
+public partial class Plugin : IPlugin, IOpAmpPlugin, IOpAmpListener<CustomMessageMessage>
 #pragma warning restore CA1515 // Consider making public types internal. Needed for AutoInstrumentation plugin loading.
 {
+    private IOpAmpClient? _opAmpClient;
+
     public void ConfigureOpAmpOptions(OpAmpClientSettings settings)
     {
         ThrowIfMissing(settings);
@@ -23,13 +26,29 @@ public partial class Plugin : IPlugin, IOpAmpPlugin
         Console.WriteLine($"{nameof(settings.MaxPendingCustomMessageBytes)}: {settings.MaxPendingCustomMessageBytes}");
     }
 
-    public void AfterOpAmpClientStarted(OpAmpClient client)
+    public void ConfigureOpAmpClient(IOpAmpClient client)
+    {
+        ThrowIfMissing(client);
+        _opAmpClient = client;
+        client.Subscribe<CustomMessageMessage>(this);
+        Console.WriteLine($"{nameof(Plugin)}.{nameof(ConfigureOpAmpClient)}() invoked.");
+    }
+
+    public void AfterOpAmpClientStarted()
     {
         Console.WriteLine($"{nameof(Plugin)}.{nameof(AfterOpAmpClientStarted)}() invoked.");
     }
 
     public void BeforeOpAmpClientStopped()
     {
+        _opAmpClient?.Unsubscribe<CustomMessageMessage>(this);
+        _opAmpClient = null;
         Console.WriteLine($"{nameof(Plugin)}.{nameof(BeforeOpAmpClientStopped)}() invoked.");
+    }
+
+    public void HandleMessage(CustomMessageMessage message)
+    {
+        ThrowIfMissing(message);
+        Console.WriteLine($"{nameof(Plugin)}.{nameof(HandleMessage)}({nameof(CustomMessageMessage)}) invoked: {message.Type}.");
     }
 }